Pure code-motion refactor (with one allowed behavior-preserving lift-extraction) breaking pkg/cluster/cluster.go (2429 LOC, 77 *Manager methods) into cohesive sibling .go files inside the existing package cluster. Per Wave-2 sibling-file rule: NOT sub-packages, NOT cluster_*.go prefix. Manager keeps its constructor, lifecycle, and event-channel ownership in a slim manager.go.
New files:
- manager.go — Manager struct, NodeState, RedundancyGroupState,
  ClusterEvent, RetryablePreFailoverError, monitorKey,
  peerGroupSnapshot, NewManager, Start/Stop, sendEvent,
  SetOnEventDrop, NodeID, ClusterID, Events, Monitor.
- heartbeat_manager.go — StartHeartbeat, StopHeartbeat,
  RestartHeartbeat, buildHeartbeat, handlePeerHeartbeat,
  handlePeerTimeout, handlePeerNeverSeen, HeartbeatStats,
  HeartbeatStats type, vrfListenConfig.
- failover.go — Single-RG + multi-RG batch manual-failover protocol
  AND all transfer-commit state machine *Locked helpers. Folds in
  the former failover_batch.go so the entire manual-failover locking
  domain (single-RG + batch + shared *Locked helpers) lives in one
  file. Includes the new
  applyTransferCommitOverridesOnPeerStateLocked helper — the only
  allowed lift-extraction, hoisting the three back-to-back override-
  application loops from the inline block formerly in
  handlePeerHeartbeat (pre-#1541 cluster.go:1516-1541) byte-for-byte.
- status.go — All Format* methods, plus InterfaceMonitorInfo,
  RethInfo, InterfacesInput types.
- peer_state.go — PeerAlive, PeerNodeID, PeerGroupStates,
  SetSoftwareVersion, SoftwareVersions, SetHAProtocolVersion,
  HAProtocolVersions, HAProtocolVersionMismatch,
  PeerMonitorStatuses.
- group_state.go — UpdateConfig, GroupStates, DataGroupIDs,
  GroupState, IsLocalPrimary, IsLocalPrimaryAny, LocalPriorities.
- readiness.go — SetRGReady.
- sync_state.go — SyncStatsProvider interface, SetSyncReady,
  IsSyncReady, SetSyncTransport, SyncTransport, SetSyncStats,
  GetSyncStats, IsSyncConnected.
- hooks.go — All Set*Hook / Set*Func methods.
- events_log.go — RecordEvent, EventHistoryFor.

Modified files:
- election.go — gained electSingleNode, SetMonitorWeight, recalcWeight.
- garp.go — gained triggerGARP (verbatim slog.Info no-op/log hook;
  native VRRP owns actual GARP transmission).

Deleted files:
- cluster.go (entire 2429 LOC distributed across sibling files).
- failover_batch.go (folded into failover.go per Gemini r2 — single
  manual-failover locking domain in one file).

Methods preserved: 64 *Manager exported methods across pkg/cluster,
all signatures identical to master. One new private helper
applyTransferCommitOverridesOnPeerStateLocked is the lift-extraction
explicitly permitted by the plan's v3 'no method bodies edited'
exception, addressing Codex r2 PLAN-NEEDS-MAJOR.

Plan + adversarial review:
- Plan doc: docs/pr/1541-cluster-mgr-split/plan.md (v3.1).
- Reviewer task IDs: docs/pr/1541-cluster-mgr-split/reviewer-ids.md.
- Codex r1 PLAN-NEEDS-MINOR; r2 PLAN-NEEDS-MAJOR; r3 PLAN-NEEDS-MINOR
  ('Fix those and I'd call this PLAN-READY'); r4 + r4-retry both
  BLOCKED-INFRA (codex-linux-sandbox ENOENT).
- Gemini r1 PLAN-KILL (transfer-commit invariant smear); r2
  PLAN-NEEDS-MINOR (acknowledged r1 fix; new findings:
  triggerGARP misplacement + failover_batch merge); r3 PLAN-READY.

Validation:
- go build ./... clean.
- go vet ./pkg/cluster/... clean.
- go test ./pkg/cluster/... — 5/5 passes (flake check).
- go test ./... — full suite, all packages pass.
- HA-sensitive PR-smoke pending: AWAITING-SMOKE / Scope:
  smoke-plus-test-failover. Smoke-runner will exercise the full
  iperf3 matrix on the loss userspace cluster plus
  'make test-failover' per repo HA validation policy.

Closes #1541.
