+++ docs/pr/1541-cluster-mgr-split/reviewer-ids.md
@@ -1,0 +1,22 @@
+# #1541 cluster manager split — reviewer task IDs
+
+## Plan review round 1 (commit dc7abd99)
+
+- Codex: `task-mpmyrii6-pl9f0f` (PLAN-NEEDS-MINOR)
+- Gemini: `task-mpmys3n8-wcaa7b` (PLAN-KILL — counter-example: transfer-commit invariant smeared across 3 files)
+
+## Plan review round 2 (commit d87d6ee5)
+
+- Codex: `task-mpmz1yi5-q7nath` (PLAN-NEEDS-MAJOR — handlePeerHeartbeat still inlined transfer-grace state)
+- Gemini: `task-mpmz2j22-zfrgrf` (PLAN-NEEDS-MINOR — acknowledged r1 resolved; 2 new findings: triggerGARP misplacement + merge failover_batch)
+
+## Plan review round 3 (commit bab60b68)
+
+- Codex: `task-mpmz7q2x-fb57wm` (PLAN-NEEDS-MINOR — doc precision)
+- Gemini: `task-mpmz87va-rr2xjb` (PLAN-READY)
+
+## Plan review round 4 — Codex doc-precision confirmation (commit c34630ec)
+
+- Codex: `task-mpmze1x5-p1svbm`
+- Gemini: skipped — Gemini already PLAN-READY at round 3 and v3.1 is documentation-only
+
